Don't clobber global 'grepprg' value
'grepprg' is a buffer-local setting. We should only use the `&l`
version, or else when we restore it after an `:Ag` command, we're
disrupting the setting in other buffers.

It's surprising that 'grepprg' is local and global, and 'grepformat' is
only global, so we have no choice but to clobber it. But this could have
it's uses: suppose you have a C program with some tests in shell (like
ag!), and you always use `ag` for 'grepprg'. The same 'grepformat' can
work project-wide, but you might have an autocommand that makes
'grepprg' use `ag --shell` for all files under the tests directory.